requirements.txt
boto3==1.40.61
botocore==1.40.61
certifi==2025.10.5
charset-normalizer==3.4.4
faiss-cpu==1.12.0
-e git+ssh://git@github.com/at15/faiss-s3.git@3e560ec9c107417124febafc4fb842c387feb940#egg=faiss_s3&subdirectory=client
filelock==3.20.0
fsspec==2025.9.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
ruff==0.14.2
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
